requirements.txt
Django==1.10.5
django-mptt==0.8.7
social-auth-app-django==1.0.0
